1.40.0-beta-20060811
git-svn-id: svn://tug.org/pdftex/trunk@65 7b5bf505-3498-49f7-8177-37ec8b90719f

+++ source/src/texk/web2c/pdftexdir/writepng.c
@@ -542,13 +542,13 @@ void write_png (integer img)
                 (checked_gamma > 1.01 || checked_gamma < 0.99))
                 tex_printf ("gamma delta=%lf ", checked_gamma);
             if (png_ptr (img)->transformations != PNG_TRANSFORM_IDENTITY)
-                tex_printf ("transform=%d", png_ptr (img)->transformations);
+                tex_printf ("transform=%lu", (long)png_ptr (img)->transformations);
             if ((png_info (img)->color_type != PNG_COLOR_TYPE_GRAY) &&
                 (png_info (img)->color_type != PNG_COLOR_TYPE_RGB) &&
                 (png_info (img)->color_type != PNG_COLOR_TYPE_PALETTE))
                 tex_printf ("colortype ");
             if (fixedpdfminorversion <= 1)
-                tex_printf ("version=%d ", fixedpdfminorversion);
+                tex_printf ("version=%d ", (int)fixedpdfminorversion);
             if (png_info (img)->interlace_type != PNG_INTERLACE_NONE)
                 tex_printf ("interlaced ");
             if (png_info (img)->bit_depth > 8)
